cron de prueba para actualizar kilometrajes de colectivos desde archivo local

+++ Dtos/ColectivoDto.cs
@@ -6,6 +6,6 @@ namespace ApiSwagger.Dtos
         public string NroColectivo { get; set; } = string.Empty;
         public string Patente { get; set; } = string.Empty;
         public string? Modelo { get; set; }
-        public int? Kilometraje { get; set; }
+    public decimal? Kilometraje { get; set; }
     }
 }
